graphql: extract exceptions
